When deleting a cart item, source of id should be the cartItem itself, not only the contained variantData.
This allows for deletion of just the cart item itself, instead of ALL INSTANCES
of cart items that have the same variantData. UseCase: Equal variants should not be
grouped together in cart, but instead be listed separately, even if the same item
exists multiple times in cart.

Note: Default behaviour is not changed. As before all items are grouped
if they are the same variant. But it allows for customizing the
cart/removeCart server function this way:

```
Meteor.default_server.method_handlers['cart/removeFromCart'] = function (cartId, cartItem) {
  check(cartId, String);
  check(cartItem, Object);
  this.unblock();

  return Cart.update({
    _id: cartId
  }, {
    $pull: {items: cartItem}
  });
};
```