PayPal Express Refund / Discount fixes (#1235)
* add check to make sure discounts are not greater than cart total

Reaction was allowing discounts greater than the cart total to be
applied, meaning orders could end up displaying an infinite amount
refunds to be paid out. This change adds client and server side checks
to make sure this doesn’t happen.

* updated surrounding div to remove extra spacing when alert is not active

There was extra spacing inside the shippingInvoice.html showing because
we were using a class on the div that had padding attached to it. This
change fixes that by using a different div class.

* updated server-side error: added Logger, changed error message

* updated error message for invalidDiscount

* updated error message for invalidDiscount

* added lodash import to shippingInvoice JS

Order details in the shippingInvoice panel were not displaying
correctly because of a missing _.lodash import. This change adds loads
back in, fixing the issue.

* fixed typo: applu -> apply

* updating err to error

* removed un-needed comment

* updating terminology

Replaced “cart” with “order” to better reflect what’s actually happening

* added invalidRefund to en.json

* updated action to not allow refunds created than original purchase amount

* rounded adjustedTotal to allow for 100% refund

adjusted total was not being rounded to a monetary-type number (i.e.
4.9899999999 instead of 4.99), so it wasn’t allowing a full refund in
some places.

* changed lodash.js to accounting.js for rounding

* fixed rounding error on discounts

* added currency symbol to refund popup

* changed source of amount to send to PayPal Express

Prior to this change, we were taking the original amount on an
authorization - which is the raw amount of your cart - and using it as
the official “capture” amount when sending a transaction to PayPal.

Because discounts are applied after the authorization, any discount
given was not sent to PayPal Express, and even though it looked like
everything was sent correctly on our UI, PayPal Express was actually
charging the full authorized amount.

This change updates the source of the amount from the original
authorization, to the final amount to be charged after all discounts.

* updated error messaging

* updated constants to match the PayPal data schema

* fixed formatting to match previous code

* created flow for 100% discounts

100% discount were failing because some payment providers see a $0.00
charge as an error ($.01 being the minimum charge).

This change introduces a flow where 100% discounts will be VOIDED when
sent to the payment provider.

* lint update to trigger PR updates

* updated lint orders.js

- triggering CI updates

* updated README.md

- add BitHound