Fix error when creating/update groups and/or group permissions for non-admin user
Resolves #3649

This PR does prevent an error being displayed when a user without the
role "shop-admin" or "owner" tries to create a new group or tries
to update permissions of an existing group.

It does so by not displaying editing buttons and having the permission
toggles not updating when switching on/off. I.e, they behave like
read-only widgets.