Partial fix for #2249.
What it does:
- Prevent unpublished handle changes to be used when assembling a link
URL to PDP.

What is does not:
- Ensure that the "not found" template is rendered in certain situations
(which I don't know what these are exactly) AFTER publishing a product.

How it does that:
If a product has unpublished changes, they published state is in
products.__published. Use that for building the perma-link from
products.__published.handle.

Adresses not the problem of the "not found"