fix: remove setting of ref
Setting ref on functional components does not work and results in a warning in React 16. With the use of recopose, and composing of HOC functional components, it becomes too dfficult to fix component by component.

Removing the `ref` prop and function definition as it may have never worked and is generally discouraged behavior.