Fix for "Reactivity error when products are not published yet" Resolves #3958
Changes:
The revision observer should push revisions to client even if product
document does not exist in session collection view (merge box) yet.
Otherwise the revision field would never be pushed to the client on
initial load of PDP, which causes the reactivity error observed in

The crux is to not "over-publish" revisions. Just adding

this.added("Products", id, fields) in line 119 would result in revisions
being pushed to the client that are completely unrelated.

This is because the current observer is not specific enough. It justs
observes all revisions that are not published yet. This is too broad.
The selector needs to be more specific and just observe revision
documents that belong to the currently published product. This fact
resulted in the changes of line 78 - 85.

Another thing where I've been wrong has to do with the comment in line
101. I think that branch is actually dead code which can be removed. I
haven't done that in this PR to keep the changes manageable. I think it
does not hurt either, but I will create a new ticket to follow up on
this.

And lastly there was a unnecessary 2nd parameter passed to `this.removed`,
which has been removed in this PR.